GettingsLibraries.py
The main function in this project is installing_libraries. The library that is not installed by this project is the colorama module, which decorates the text, but if it is not present, the program itself will try to install it, if it fails, an error will be displayed. The getting_libraries function is used in the main function and installs the libraries. Using multithreading, the function works with the local function letter_animation, which outputs animated text, as well as with the local function subprocess_output, which actually installs the necessary libraries.